move get_avg_point_color to image_utils
